+++ build.sbt
@@ -21,5 +21,6 @@ lazy val root = (project in file(".")).
       "-Xfatal-warnings",  // New lines for each options
       "-deprecation",
       "-feature",
+      "-unchecked"
     )
   )
